Get new structure working on windows systems
